Alphabet horizontal swiping was framed as reusable HOC. AlphabetSelector component refactored
